[REFACTOR] Layout컴포넌트 분리 - Headere, Footer컴포넌트 생성 후 Layout페이지에서 분리 - device의 넓이가 800이상 ( tablet ) 일 때부터 header의 아이콘 추가 - 햄버거 버튼의 메뉴 보기에서 나열된 메뉴보기로 반응형 디자인 추가 - Header에서 나열된 메뉴 버튼 컴포넌트를 따로 분리함 ->MenuButton.tsx생성: 각 메뉴의 리스트 링크와 Icon이 있는 컴포넌트가 됨
