+++ src/shared/Layout/styles/header.module.css
@@ -1,4 +1,3 @@
-/* header */
 .header {
   position: sticky;
   top: 0;
@@ -22,14 +21,6 @@
   cursor: pointer;
 }
 
-/* footer */
-.footer {
-  padding: 2rem;
-  text-align: center;
-  border-top: 2px solid var(--color-border);
-  color: var(--color-muted);
-}
+@media (min-width: 800px) {
 
-.footer a {
-  color: var(--color-muted);
 }
